adding in the set_masked_values_to_zero
