+++ pose_evaluation/utils/test_pose_utils.py
@@ -14,11 +14,12 @@ from pose_evaluation.utils.pose_utils import (
     copy_pose,
     get_face_and_hands_from_pose,
     reduce_pose_components_and_points_to_intersection,
-    preprocess_pose,
+    # preprocess_pose,
     get_component_names_and_points_dict,
-    preprocess_poses,
+    # preprocess_poses,
     detect_format,
     zero_pad_shorter_poses,
+    set_masked_to_origin_position
 )
 
 
@@ -220,48 +221,77 @@ def test_detect_format(
             detect_format(pose)
 
 
-def test_preprocess_pose(test_mediapipe_poses_paths: List[Path]):
-    poses = [load_pose_file(pose_path) for pose_path in test_mediapipe_poses_paths]
-    preprocessed_poses = []
+# def test_preprocess_pose(test_mediapipe_poses_paths: List[Path]):
+#     poses = [load_pose_file(pose_path) for pose_path in test_mediapipe_poses_paths]
+#     preprocessed_poses = []
 
-    data_arrays = [pose.body.data for pose in poses]
+#     data_arrays = [pose.body.data for pose in poses]
 
-    for pose in poses:
-        processed_pose = preprocess_pose(pose,
-                        normalize_poses=True,                        
-                        remove_legs=True,
-                        remove_world_landmarks=True,
-                        conf_threshold_to_drop_points=0.2)
-        #TODO: check expected result
+#     for pose in poses:
+#         processed_pose = preprocess_pose(pose,
+#                         normalize_poses=True,                        
+#                         remove_legs=True,
+#                         remove_world_landmarks=True,
+#                         conf_threshold_to_drop_points=0.2)
+#         #TODO: check expected result
 
 
-def test_preprocess_poses(test_mediapipe_poses: List[Pose]):
+# def test_preprocess_poses(test_mediapipe_poses: List[Pose]):
 
-    nan_counts = [np.count_nonzero(np.isnan(pose.body.data)) for pose in test_mediapipe_poses]
+#     nan_counts = [np.count_nonzero(np.isnan(pose.body.data)) for pose in test_mediapipe_poses]
 
-    preprocessed_poses = preprocess_poses(
-        test_mediapipe_poses,
-        normalize_poses=True,
-        reduce_poses_to_common_points=True,
-        remove_world_landmarks=True,
-        remove_legs=True,
-        zero_pad_shorter_pose=True,
-        conf_threshold_to_drop_points=0.2,
-    )
+#     preprocessed_poses = preprocess_poses(
+#         test_mediapipe_poses,
+#         normalize_poses=True,
+#         reduce_poses_to_common_points=True,
+#         remove_world_landmarks=True,
+#         remove_legs=True,
+#         zero_pad_shorter_pose=True,
+#         conf_threshold_to_drop_points=0.2,
+#     )
 
-    for i, pose in enumerate(preprocessed_poses):
-        component_names, points_dict = get_component_names_and_points_dict(pose)
-        assert "LEFT_KNEE" not in points_dict["POSE_LANDMARKS"]
-        assert "POSE_WORLD_LANDMARKS" not in component_names
+#     for i, pose in enumerate(preprocessed_poses):
+#         component_names, points_dict = get_component_names_and_points_dict(pose)
+#         assert "LEFT_KNEE" not in points_dict["POSE_LANDMARKS"]
+#         assert "POSE_WORLD_LANDMARKS" not in component_names
 
-        # zero-padded properly? Should all be the same number of frames
-        assert pose.body.data.shape[0] == preprocessed_poses[0].body.data.shape[0]
+#         # zero-padded properly? Should all be the same number of frames
+#         assert pose.body.data.shape[0] == preprocessed_poses[0].body.data.shape[0]
 
-        # do we have nan values?
-        nan_count = np.count_nonzero(np.isnan(pose.body.data))
-        assert np.count_nonzero(np.isnan(pose.body.data)) == nan_counts[i]
-        assert nan_count == 0
+#         # do we have nan values?
+#         nan_count = np.count_nonzero(np.isnan(pose.body.data))
+#         assert np.count_nonzero(np.isnan(pose.body.data)) == nan_counts[i]
+#         assert nan_count == 0
         
+
+def test_set_masked_to_origin_pos(test_mediapipe_poses: List[Pose]):
+    # Create a copy of the original poses for comparison
+    originals = [copy_pose(pose) for pose in test_mediapipe_poses]
+
+    # Apply the transformation
+    poses = [set_masked_to_origin_position(pose) for pose in test_mediapipe_poses]
+
+    for original, transformed in zip(originals, poses):
+        # 1. Ensure the transformed data is still a MaskedArray
+        assert isinstance(transformed.body.data, np.ma.MaskedArray)
+
+        # 2. Ensure the mask is now all False
+        assert np.all(transformed.body.data.mask == False)
+
+        # 3. Check the shape matches the original
+        assert transformed.body.data.shape == original.body.data.shape
+
+        # 4. Validate masked positions in the original are now zeros
+        assert np.all(
+            transformed.body.data.data[original.body.data.mask] == 0
+        )
+
+        # 5. Validate unmasked positions in the original remain unchanged
+        assert np.all(
+            transformed.body.data.data[~original.body.data.mask]
+            == original.body.data.data[~original.body.data.mask]
+        )
+
 
 
 def test_hide_low_conf(test_mediapipe_poses: List[Pose]):
